Organise Testing and Linting documentation
It seems sensible to gather all our tests and
linting references in one place that signposts to relevant locations
(it can then be referred to in contributing and readme) and tells
user how to run all tests at once.

This moves “Testing components on their own” and “Updating component
snapshots” to `/coding-standards/components.md`

Remove `npm test` from “Javascript” as this is documented in
`testing-linting.md`.