Improve readability of links
Reduce the thickness of link underlines and offset them from the text in order to improve readability. This should help especially when multiple links are included in long lists (for example, search results pages) or at larger sizes.

Allow the thickness and offset to be overridden, so that anyone using GOV.UK Frontend outside of GOV.UK can opt out of these changes.

Co-authored-by: Chris Thomas <christopher.thomas@digital.cabinet-office.gov.uk>
Co-authored-by: Charlotte Downs <charlotte.downs@digital.cabinet-office.gov.uk>