Add inverse link mixin and modifier class
We’ll need this for the header, which has white links on a black background. It’s also a fairly common thing for service teams to need to implement, and we often see issues with the focus state when it’s implemented badly, so it’s worth making this reusable.