Add no-underline mixin and modifier class
This is a fairly common thing for service teams to need to implement, and they’ll often do it using `text-decoration` which, as a shorthand property, results in the `text-decoration-thickness` being reset to auto for the hover state.

Provide a mixin and class which includes the correct behaviour. Add it to the ‘links’ examples page.