Only bind to localhost by default
By default, Browsersync will listen on all interfaces.

In line with security best-practises, default to only listening on localhost, which means by default we’d only accept connections from the same machine.

We have cases where we may want to allow users to connect to the review app externally (e.g. testing on a mobile device whilst connected to the same LAN as a developer machine), so preserve the existing behaviour if the environment variable `ALLOW_EXTERNAL_CONNECTIONS` is `”true”`.

Add a note to the ‘running locally’ doc to explain this.